chore: prefix internal files and add deprecation messages
